# src/main/resources/apidef/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a demo online store.",
    "version": "1.0.0",
    "title": "Product Store",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "scott.seo@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "dvdstore.scottwseo.com",
  "basePath": "/api/v1/dvdstore",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/products": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Lists products",
        "description": "Returns a list of products",
        "operationId": "listProducts",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "description": "Starting index of page",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "size",
            "in": "query",
            "description": "Size of array",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Products"
            }
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "products"
        ],
        "summary": "Add a new product to the store",
        "description": "",
        "operationId": "addProduct",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Product object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductCreate"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "dvdstore_auth": [
              "write:products",
              "read:products"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "products"
        ],
        "summary": "Update an existing product",
        "description": "",
        "operationId": "updateProduct",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Product object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductCreate"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Product not found"
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "dvdstore_auth": [
              "write:products",
              "read:products"
            ]
          }
        ]
      }
    },
    "/product/{productId}": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Find product by ID",
        "description": "Returns a single product",
        "operationId": "getProductById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ProductCreate"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "products"
        ],
        "summary": "Deletes a product",
        "description": "",
        "operationId": "deleteProduct",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "productId",
            "in": "path",
            "description": "Product id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "dvdstore_auth": [
              "write:products",
              "read:products"
            ]
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "dvdstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://productstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:products": "modify products in your account",
        "read:products": "read your products"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "Pagination": {
      "properties": {
        "start": {
          "type": "integer",
          "format": "int64",
          "description": "requested starting index of page"
        },
        "resultSize": {
          "type": "integer",
          "format": "int64",
          "description": "number of items per page"
        },
        "next": {
          "type": "string",
          "format": "url",
          "description": "url to next page"
        },
        "prev": {
          "type": "string",
          "format": "url",
          "description": "url to previous page"
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "description": "total number of items"
        },
        "totalPages": {
          "type": "integer",
          "format": "int64",
          "description": "total number of pages"
        }
      }
    },
    "Products": {
      "properties": {
        "pagination": {
          "$ref": "#/definitions/Pagination"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Product"
          }
        }
      }
    },
    "ProductCreate": {
      "type": "object",
      "required": [
        "category",
        "title",
        "actor",
        "price",
        "common_prod_id"
      ],
      "properties": {
        "prod_id": {
          "type": "integer",
          "format": "int64"
        },
        "category": {
          "type": "integer",
          "format": "int64",
          "example": "1"
        },
        "title": {
          "type": "string",
          "example": "star wars"
        },
        "actor": {
          "type": "string",
          "example": "harrison ford"
        },
        "price": {
          "type": "number",
          "example": "10.99"
        },
        "common_prod_id": {
          "type": "integer",
          "format": "int64",
          "example": "1"
        }
      }
    },
    "Product": {
      "type": "object",
      "required": [
        "category",
        "title",
        "actor",
        "price",
        "common_prod_id"
      ],
      "properties": {
        "prod_id": {
          "type": "integer",
          "format": "int64"
        },
        "category": {
          "type": "string",
          "example": "Action"
        },
        "title": {
          "type": "string",
          "example": "star wars"
        },
        "actor": {
          "type": "string",
          "example": "harrison ford"
        },
        "price": {
          "type": "number",
          "example": "10.99"
        },
        "common_prod_id": {
          "type": "integer",
          "format": "int64",
          "example": "1"
        }
      }
    },
...
  },
...
}